refactor: move mocket.mockredis to mocket.compat.mockredis
